// src/models/NounCategory.ts
/* eslint-disable */
// enum for noun categories like things that can be drawn, things that can be moved, things that can be eaten, living things
export enum NounCategory {
  Drawable = "drawable",
  Movable = "movable",
  Edible = "edible",
  Cookable = "cookable",
  Living = "living",
  Person = "person",
  Material = "material",
  Noisy = "noisy",
  Property = "property",
  Process = "process",
}